+++ types/tgl_peer_id.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstdint>
+
 enum class tgl_peer_type {
     unknown = 0,
     user = 1,
@@ -13,10 +15,10 @@ enum class tgl_peer_type {
 
 struct tgl_peer_id_t {
     tgl_peer_type peer_type;
-    int peer_id;
-    long long access_hash;
+    int32_t peer_id;
+    int64_t access_hash;
     tgl_peer_id_t() : peer_type(tgl_peer_type::unknown), peer_id(0), access_hash(0) {}
-    tgl_peer_id_t(tgl_peer_type peer_type, int peer_id, long long access_hash = 0)
+    tgl_peer_id_t(tgl_peer_type peer_type, int32_t peer_id, int64_t access_hash = 0)
         : peer_type(peer_type), peer_id(peer_id), access_hash(access_hash) {}
 };
 
@@ -26,17 +28,17 @@ inline bool operator==(const tgl_peer_id_t& lhs, const tgl_peer_id_t& rhs)
 }
 
 struct tgl_peer_id_user : public tgl_peer_id_t {
-    tgl_peer_id_user(int user_id) : tgl_peer_id_t(tgl_peer_type::user, user_id) {}
+    tgl_peer_id_user(int32_t user_id) : tgl_peer_id_t(tgl_peer_type::user, user_id) {}
 };
 
 struct tgl_peer_id_chat : public tgl_peer_id_t {
-    tgl_peer_id_chat(int chat_id) : tgl_peer_id_t(tgl_peer_type::chat, chat_id) {}
+    tgl_peer_id_chat(int32_t chat_id) : tgl_peer_id_t(tgl_peer_type::chat, chat_id) {}
 };
 
 struct tgl_peer_id_channel : public tgl_peer_id_t {
-    tgl_peer_id_channel(int channel_id) : tgl_peer_id_t(tgl_peer_type::channel, channel_id) {}
+    tgl_peer_id_channel(int32_t channel_id) : tgl_peer_id_t(tgl_peer_type::channel, channel_id) {}
 };
 
 struct tgl_peer_id_enc_chat : public tgl_peer_id_t {
-    tgl_peer_id_enc_chat(int enc_chat_id) : tgl_peer_id_t(tgl_peer_type::enc_chat, enc_chat_id) {}
+    tgl_peer_id_enc_chat(int32_t enc_chat_id) : tgl_peer_id_t(tgl_peer_type::enc_chat, enc_chat_id) {}
 };
